+++ CHANGELOG.md
@@ -7,13 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.2] 2026-05-08
+
 ### Added 
 
 - Added `get_degree_distribution` to load degree distribution from QC files.
+- `get_hash_stats` now collects sample confidence data.
+- Metrics `Number of cells` and `% Sample called cells` have been added to the Key Metrics Table.
+- Component "Component sample confidence" added to the Quality Metrics tab.
 
 ### Changed
 
 - Added new slots to example data jsons with degree distribution. 
+- .pxl files are now tracked with Git LFS.
+- `get_test_data` now loads data from `extdata` to allow easier update of test data and more representative test data.
+- `get_test_data` now creates a temporary folder where it puts the .pxl files to be read. This avoids `Duckdb` file accession collisions in CI.
+- The Hash Purity plot's y-axis now ranges 90-100% by default.
 
 ## [0.9.1] 2026-04-30